requirements.txt
acora==2.2
adblockparser==0.7
AdvancedHTTPServer==2.2.0
aiocmd==0.1.2
aioconsole==0.3.1
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.5.3
aiomultiprocess==0.8.0
aiorpcX==0.18.5
aiosqlite==0.15.0
aiowinreg==0.0.3
ajpy==0.0.4
alembic==1.4.3.dev0
aniso8601==8.1.1
apispec==3.3.1
apispec-webframeworks==0.5.2
apparmor==2.13.6
appdirs==1.4.4
argcomplete==1.8.1
arrow==0.17.0
artifacts==20201106
asciitree==0.3.3
asgiref==3.3.0
asn1crypto==1.4.0
astroid==2.5
async-timeout==3.0.1
asysocks==0.0.2
attrs==20.3.0
autobahn==17.10.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backdoor-factory==0.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
binwalk==2.2.0
bitstring==3.1.7
blinker==1.4
boltons==19.1.0
bottle==0.12.19
Brotli==1.0.9
cairocffi==1.2.0
capstone==4.0.2
cbor==1.0.0
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.4
Chameleon==3.8.1
chardet==4.0.0
CherryPy==8.9.1
chirp==0.3.0.dev0
click==7.1.2
click-plugins==1.1.1
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
crackmapexec==5.1.4.dev0
crcelk==1.3
cryptography==3.3.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.11
deprecation==2.0.7
dhcpig==0.0.0
dicttoxml==1.7.4
distro==1.5.0
dj-database-url==0.5.0
Django==2.2.18
django-heroku==0.3.1
djangorestframework==3.12.2
dnslib==0.9.14
dnspython==2.0.0
EasyProcess==0.2.5
ecdsa==0.16.1
Electrum==4.0.9
email-validator==1.1.1
ExifRead==2.3.2
expiringdict==1.1.4
faraday-client==1.0.0
faraday-plugins==1.2
faradaysec==3.14.0
feedparser==5.2.1
fierce==1.4.0
filedepot==0.5.2
filteralchemy==0.1.0
flasgger==0.9.5
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Classful==0.14.1
Flask-HTTPAuth==3.2.4
Flask-KVSession-fork==0.6.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Restless==0.17.0
Flask-Security-Too==4.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.18.2
fuzzywuzzy==0.18.0
GDAL==3.2.1
GeoIP==1.3.2
geoip2==2.9.0
geojson==2.5.0
gevent==20.9.0
git-cola==3.8
goofile==1.6
google==2.0.3
gpg===1.14.0-unknown
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.17
grequests==0.4.0
gunicorn==20.0.4
gvm-tools==21.1.0
gyp==0.1
h11==0.12.0
h2==4.0.0
hashID==3.1.4
helpdev==0.7.1
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
httplib2==0.18.1
humanize==0.0.0
hupper==1.10.2
hyperframe==6.0.0
hyperlink==19.0.0
icalendar==4.0.3
idna==2.10
imageio==2.4.1
impacket==0.9.22
importlib-metadata==1.6.0
incremental==17.5.0
intervaltree==3.0.2
invoke==1.4.1
ipdb==0.13.3
ipwhois==0.15.1
IPy==1.1
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jsonrpclib-pelix==0.4.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kaitaistruct==0.9
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kismetexternal==2019.12.1
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
ldap3==2.8.1
ldapdomaindump==0.9.3
LibAppArmor==2.13.6
logutils==0.3.3
lsassy==2.1.2
lxml==4.6.2
lz4==3.1.3+dfsg
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.19.0
mate-menu==20.4.1
mate-tweak==21.4.1
matplotlib==3.3.4
maxminddb==2.0.3
mccabe==0.6.1
mechanize==0.4.5
meld==3.20.2
mercurial==5.6.1
mimerender==0.6.0
minidump==0.0.12
minikerberos==0.2.0
mistune==0.8.4
mitmproxy==6.0.2
mock==4.0.3
more-itertools==4.2.0
msgpack==1.0.0
msldap==0.2.10
multidict==5.1.0
mutagen==1.45.1
mysqlclient==1.4.4
nassl==4.0.0
nbformat==5.1.2
neo4j==1.7.0.dev0
neobolt==1.7.17
neotime==1.7.4
netaddr==0.7.19
netifaces==0.10.9
networkx==2.5
nose==1.3.7
nplusone==1.0.0
ntlm-auth==1.4.0
numpy==1.19.4
nyx==2.1.0
olefile==0.46
onboard==1.4.1
onioncircuits==0.7
onionshare==2.2
ospd==20.8.2
ospd-openvas==20.8.1
packaging==20.9
paramiko==2.7.2
parsedatetime==2.6
parsero==0.82
parso==0.8.1
passlib==1.7.2
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==2.0.2
patator==0.9
pbkdf2==1.3
pbr==5.5.0
pcapy==0.11.4
pefile==2019.4.18
pexpect==4.8.0
pgspecial==1.11.10
phonenumbers==8.12.1
pickleshare==0.7.5
Pillow==8.1.0
plaster==1.0
plaster-pastedeploy==0.5
plecost==1.1.2
plotly==4.14.3
pluginbase==1.0.0
ply==3.11
prettytable==0.7.2
promise==2.2
prompt-toolkit==3.0.14
protobuf==3.12.4
psutil==5.8.0
psycopg2==2.8.6
publicsuffix2==2.20191221
py-ubjson==0.16.1
pyaes==1.6.1
pyaff4==0.26.post6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycares==3.1.1
pycountry==20.7.3
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.7
pycups==2.0.1
pycurl==7.43.0.6
pydot==1.4.1
pyee==7.0.2
pyelftools==0.27
pyenchant==3.2.0
pygeoip==0.3.2
Pygments==2.7.1
PyGObject==3.38.0
pygraphviz==1.6
PyHamcrest==1.9.0
PyICU==2.5
pyinotify==0.9.6
pylint==2.7.1
pylnk3==0.2.1
pymongo==3.11.0
pymssql==2.1.4
PyNaCl==1.4.0
pynvim==0.4.2
pyodbc==4.0.30
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyotp==2.3.0
pyparsing==2.4.7
pypcapfile==0.12.1
PyPDF2==1.26.0
pyperclip==1.8.0
pypng==0.0.20
pyppeteer==0.2.5
pypsrp==0.5.0
pypykatz==0.3.7
PyQRCode==1.2.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.1
pyramid==1.10.4
PyRIC==0.1.6.4
pyrsistent==0.15.5
pyrtlsdr==0.2.91
pyserial==3.5b0
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pyspnego==0.1.5
python-apt==2.1.7
python-dateutil==2.8.1
python-debian==0.1.39
python-debianbts==3.1.0
python-docx==0.8.10
python-editor==1.0.3
python-gvm==21.1.1
python-Levenshtein==0.12.1
python-libnmap==0.7.2
python-magic==0.4.20
python-magic-ahupp==0.4.13
python-memcached==1.58
python-mimeparse==1.6.0
python-pam==1.8.4
python-pptx==0.6.18
python-snappy==0.5.3
python-xapp==2.0.2
python-xlib==0.29
PyTrie==0.2
pytsk3==20200117
pytz==2021.1
pyusb==1.0.2
PyVirtualDisplay==0.2.1
pywerview==0.2.0
pywinrm==0.4.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
pyzmq==20.0.0
QDarkStyle==2.8.1
qrcode==6.1
QtPy==1.9.0
rdflib==5.0.0
redis==3.5.3
rekall-core==1.7.2rc1
rekall-lib==1.7.1
reportbug==7.9.0
repoze.lru==0.7
requests==2.25.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
retrying==1.3.3
rfcat==1.0.1
roguehostapd==1.1.2
Routes==2.5.1
rq==1.7.0
rtlsdr-scanner==1.3.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
rule-engine==1.1.0
Rx==3.1.1
scapy==2.4.4
scipy==1.6.0
secure==0.2.1
selenium==4.0.0a1
service-identity==18.1.0
setproctitle==1.2.1
shodan==1.25.0
simplejson==3.17.2
simplekv==0.13.0
six==1.15.0
smoke-zephyr==2.0.1
sortedcontainers==2.1.0
soupsieve==2.1
SPARQLWrapper==1.8.5
speaklater==1.4
specan==0.0.0
SQLAlchemy==1.3.22
sqlalchemy-schemadisplay==1.3
sqlparse==0.4.1
sslyze==4.0.3
stem==1.8.0
syslog-rfc5424-formatter==1.2.2
tabulate==0.8.7
Tempita==0.5.2
termcolor==1.1.0
terminaltables==3.1.0
termineter==1.0.4
texttable==1.6.3
theHarvester==3.2.3
tld==0.11.11
tls-parser==1.2.2
toml==0.10.2
torbrowser-launcher==0.3.3
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
translationstring==1.4
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
u-msgpack-python==2.3.0
ufw==0.36
unicodecsv==0.14.1
Unidecode==1.2.0
urllib3==1.26.2
urwid==2.1.2
uvloop==0.14.0
venusian==3.0.0
vinetto==0.8.0
visvis==1.11.1
wafw00f==2.1.0
waitress==1.4.4
wapiti3==3.0.3
wcwidth==0.1.9
webargs==7.0.1
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.57.0
websockets==8.1
websploit==4.0.4
WebTest==2.0.35
Werkzeug==1.0.1
wfuzz==3.1.0
whitenoise==5.2.0
whois==0.8
wifiphisher==1.4
wifite==2.5.8
winacl==0.0.2
wrapt==1.12.1
wsaccel==0.6.2
wsproto==0.15.0
WTForms==2.2.1
wxPython==4.0.7
xcffib==0.8.1
xdg==5
XlsxWriter==1.1.2
xmltodict==0.12.0
yara-python==4.0.2
yarl==1.6.3
yaswfp==0.9.3
youtube-dl==2021.2.4.1
zipp==1.0.0
zope.component==4.3.0
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
